Fix spells the DDA version broke
Updated Magic Signs in the DDA version to use mana instead of fatigue.

Caused by https://github.com/CleverRaven/Cataclysm-DDA/pull/47852
existing basically solely to fuck over modders, because evidently if a
feature is underused in-repo the answer is to axe it instead of find a
use for it.

Of course Arcana being the only thing that used it probably added some
encouragment to remove it.