Belated fix times for various powerup items
Set it so that the various tools that grant a buff through consume_drug action have increased duration. The old durations seem to be a holdover from

The comestibles like potions meanwhile are going to need a more involved look through to see what their durations should be since their current durations vary a lot more, so I'll tackle that separately.