Weapon expansions for the two martial arts
* Added the monomolecular blade to the weapon list for Via Gladium et
Malleo. It's a one-handed cutting weapon with over 20 damage, but the
main reasoning is this prevents Noct's Arcane Operative profession from
being the only Cleansing Flame-affiliated profession unable to make use
of the style with their starting gear.
* Added silver athames, moonstone fangs, and a limited selection of
vanilla knifes to the weapon list for Shrike's Talons. As with the above
change, main reasoning is this ensues standard blood mages have a
hypothetical option to make use of the style should they start with it.