More misc updates
* Finally fixed up interrmittent activation effects and depeciated use
of AEP_EVIL in the BN version.
* Also re-implemented moonstone fang's effect on healthiness and sanity
as a potential on-hit effect instead.
* Overhauled books spawns a bit, now properly dividing up itemgroups for
locations that were unlikely to have some of the last books that were
produced (namely A Story In Shadow and the magitech lab jounral), with
the standard "spawn anything relevant" itemmgroups now confined to more
logical locations.