Prepartion for itemgroup overhaul in DDA version
Setting aside a branch to have the override itemgroups in the DDA version already prepped for if https://github.com/CleverRaven/Cataclysm-DDA/pull/56469 is merged. Gonna have to do similar for other mods I'm involved in maintaining, but this is a good start for getting used to the new format (also easier as I'm fairly certain this is the only mod I'm involved in that has vanilla-override itemgroups actually set aside separate from new groups).