Added longest_side to crossbows which missed it
For grand crossbow it's from original's huge grossbow, 102 cm. Not sure if «copy from» is better here.
For pistol - in DDA pistol crossbow is 70 cm which is inadequate, will post issue there or will try to change it myself. Now 45 cm to be able to store it in XL holsters.