Profession update
* Added a new profession for use in the advanced arcanist scenarios, in
the same vein as arcane purifiers/operatives and sanguine shrikes, being
a faction arcanist better geared for post-cata scenarios. In this case
it's a Keeper arcanist that's in some ways comparable to The Hermit in
being driven to wander, albeit balanced to be an advanced profession.
* Just as the other advanced arcanist professions, the new profession
also gets a specific chargen-only spell, this one geared for making a
quick escape from melee.
* Also finally realized that I can change profession IDs without
breaking old saves so I updated something I'd changed the display name
of years ago.
* Mucked about with proficiencies for professions in the DDA version a
little bit more too.