Split up arcanist house and basement a bit
Added a second variant of the arcanist house with a different appearance
and layout, and dividied up the arcanist basement variations between the
two. Both to give a bit more variety, and so there won't be a visible
Cleansing Flame raid on a house that has a Cleansing Flame hideout
underneath it.