Alteration to missiondefs
Due to problems that ensue if the player has viewed most of the surrounding area, I've gone ahead and decided to set it up so it doesn't enforce unseen map areas for missions, even if there's the risk it'll eat a patch of forest the player has been near.

Also made it so the mapgen updates triggered from the refugee center and other locations SHOULD be much more robust.

Also spotted obsolete traits on Johnathan.