Destructive Absorption updates
Fixed the duration of Destructive Absorption's aftereffect to be consistent across both versions, and to also apply roughly the same stamina and healthiness across a duration that's now actually scaled as DoT/paralysis effects are. Also set it so it counts as true/pure damage, fitting since it's the top-tier Dragonblood spell, compare with Wave of Destruction.