Add another magitech CBM
Added kinetic amplifier CBM, as a magictech bionic melee weapon. Serves as a blunt-focused unarmed weapon, with a crit tec that adds high knockback, damage, and armor penetration. As it also warns however, it also has an on-hit effect that goes off like a flashbang when it triggers. Potentially exploitable by a flash-protected cyborg (or someone with a mask of insight, or a sanguine shrike using Harden Senses, etc), otherwise a risk that makes it a double-edged sword.

Also adds salvaged version to the Arcana/Cata++ patchmod.