Rebalances to inlay gunmods, arcane purifier
Tweaked the relative power and efficiency, along with the misc drawbacks, of the inlay gunmods along with removing that janky-as-fuck INSTALL_DIFFICULT flag. Fission inlay should now be a bit more worth the risk of dimensional fatigue without needing to add bigger boom since it gets a direct damage modifier.

And to once and for all solve the unending problems with lasers fitting in holsters, gave the arcane purifier a M1911 with the electro anomaly inlay installed. That should solve things until DDA finds some way to make a 1911 fail to fit in a regular holster.

Fixes https://github.com/chaosvolt/cdda-arcana-mod/issues/312