Impact site tweaks
Fixed it being the only one of the four to not start with any accessible water, and fixed the slabs being able to generate roof overhead.

I actually tested making them a legit crater and making it so access to the temple occurs solely via an elevator-like platform, but it's too janky in practice.