Rework air talisman effect
Buff updated to account for immunity flags not applying to monsters,  the main intended recipient of the effect. Immunities expanded a bit since those do correctly seem to apply to monsters, and some base_mods applied to add bonuses to affected monsters. Now applies the buff to any friendly, not just summoned monsters.

Since applying the pet effect to temp summoned monsters is a bit weird and hacky, it focuses on just the buff and the AoE heal.